% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{html_letter}
\alias{html_letter}
\title{Create a letter in HTML}
\usage{
html_letter(..., css = c("default", "letter"))
}
\arguments{
\item{..., css}{Arguments passed to \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is similar to \code{html_paged}. The only difference is in
the default stylesheets. See \url{https://pagedown.rbind.io/html-letter/} for
an example.
}
