% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{html_propuesta}
\alias{html_propuesta}
\title{Propuesta con formato de ProyAIS}
\usage{
html_propuesta(
  ...,
  css = c("default", "letter", "propuesta"),
  fig_caption = FALSE
)
}
\value{
An R Markdown output format.
}
\description{
Propuesta con formato de ProyAIS
}
