% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{html_resume}
\alias{html_resume}
\title{Create a resume in HTML}
\usage{
html_resume(..., css = "resume", template = pkg_resource("html", "resume.html"))
}
\arguments{
\item{..., css, template}{See \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is based on Min-Zhong Lu's HTML/CSS in the Github repo
\url{https://github.com/mnjul/html-resume}.
}
\references{
\url{https://rstudio.github.io/pagedown/html-resume.html}
}
