% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{jss_paged}
\alias{jss_paged}
\title{Create an article for the Journal of Statistical Software}
\usage{
jss_paged(..., css = c("jss-fonts", "jss-page", "jss"), template = pkg_resource("html", 
    "jss_paged.html"), highlight = NULL, pandoc_args = NULL, 
    base_format = rmarkdown::html_document)
}
\arguments{
\item{..., css, template, highlight, pandoc_args, base_format}{Arguments passed to \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is similar to \code{\link{html_paged}}.
}
