% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{knit_pdf_chrome}
\alias{knit_pdf_chrome}
\title{Convert Rmd to PDF with Chrome}
\usage{
knit_pdf_chrome(input, encoding)
}
\arguments{
\item{input}{The input file to be rendered. This can be an R script (.R),
an R Markdown document (.Rmd), or a plain markdown document.}

\item{encoding}{Not used. This argument is required by
RStudio IDE.}
}
\value{
Path of the output file (invisibly).
}
\description{
This is a helper function intended to be used with RStudio IDE. When the
line \code{"knit: knit_pdf_chrome"} is present in the YAML header of an Rmd
file, the behavior of the 'Knit' button of the RStudio IDE is modified: an
extra PDF file is also generated using Chrome. This function is suitable for
any R Markdown HTML output formats.
}
\details{
If \code{knit_pdf_chrome()} cannot automatically find Chrome, use the
environment variable \code{PAGEDOWN_CHROME} to set the path to Google Chrome
or Chromium executable.
}
\seealso{
\link{chrome_print}()
}
\keyword{internal}
