% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{mdpi_paged}
\alias{mdpi_paged}
\title{Create a manuscript for for submission to MDPI}
\usage{
mdpi_paged(..., css = c("default", "mdpi"), template = pkg_resource("html", 
    "mdpi-paged.html"), csl = pkg_resource("csl", "mdpi.csl"), highlight = NULL, 
    pandoc_args = NULL, toc = FALSE)
}
\arguments{
\item{..., css, template, csl, highlight, pandoc_args}{Arguments passed to \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is similar to \code{\link{html_paged}}.
}
