% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagedown-package.R
\docType{package}
\name{pagedown-package}
\alias{pagedown}
\alias{pagedown-package}
\title{pagedown: Paginate the HTML Output of R Markdown with CSS for Print}
\description{
Use the paged media properties in CSS and the JavaScript library 'paged.js' to split the content of an HTML document into discrete pages. Each page can have its page size, page numbers, margin boxes, and running headers, etc. Applications of this package include books, letters, reports, papers, business cards, resumes, and posters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rstudio/pagedown}
  \item Report bugs at \url{https://github.com/rstudio/pagedown/issues}
}

}
\keyword{internal}
