% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poster.R
\name{poster_relaxed}
\alias{poster_relaxed}
\title{Create posters in HTML}
\usage{
poster_relaxed(..., css = "poster-relaxed", template = pkg_resource("html", 
    "poster-relaxed.html"), number_sections = FALSE)
}
\arguments{
\item{..., css, template, number_sections}{See \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
The output format \code{poster_relaxed()} is based on an example in the
Github repo \url{https://github.com/RelaxedJS/ReLaXed-examples}. See
\url{https://pagedown.rbind.io/poster-relaxed/} for an example.
}
