% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{book_crc}
\alias{book_crc}
\title{Create a book for Chapman & Hall/CRC}
\usage{
book_crc(..., css = c("crc-page", "default-page", "default", "crc"))
}
\arguments{
\item{..., css}{Arguments passed to \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is similar to \code{\link{html_paged}}. The only
difference is in the default stylesheets.
}
