% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{business_card}
\alias{business_card}
\title{Create business cards}
\usage{
business_card(template = pkg_resource("html", "card.html"))
}
\arguments{
\item{template}{The path to the Pandoc template to convert Markdown to HTML.}
}
\value{
An R Markdown output format.
}
\description{
This output format is based on an example in the Github repo
\url{https://github.com/RelaxedJS/ReLaXed-examples}. See
\url{https://pagedown.rbind.io/business-card/} for an example.
}
\examples{
pagedown::business_card()
}
