% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{chrome_print}
\alias{chrome_print}
\title{Print a web page to PDF using the headless Chrome}
\usage{
chrome_print(url, output = xfun::with_ext(url, "pdf"), browser = "google-chrome", 
    wait = 5, extra_args = c("--disable-gpu"), verbose = FALSE)
}
\arguments{
\item{url}{A URL or local file path to a web page.}

\item{output}{The (PDF) output filename. For a local web page
\file{foo/bar.html}, the default PDF output is \file{foo/bar.pdf}; for a
remote URL \file{https://www.example.org/foo/bar.html}, the default output
will be \file{bar.pdf} under the current working directory.}

\item{browser}{Path to Google Chrome or Chromium. This function will try to
find it automatically if the path is not explicitly provided.}

\item{wait}{The number of seconds to wait before deeming the page to be
ready. Use a larger value if the page takes longer to load completely.}

\item{extra_args}{Extra command-line arguments to be passed to Chrome.}

\item{verbose}{Whether to show verbose command-line output.}
}
\value{
Path of the output file (invisibly).
}
\description{
This is a wrapper function to execute the command \command{chrome --headless
--print-to-pdf url}. Google Chrome (or Chromium on Linux) must be installed
prior to using this function.
}
\references{
\url{https://developers.google.com/web/updates/2017/04/headless-chrome}
}
