% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrome.R
\name{chrome_print}
\alias{chrome_print}
\title{Print a web page to PDF using the headless Chrome (experimental)}
\usage{
chrome_print(url, output = xfun::with_ext(url, "pdf"), browser = "google-chrome", 
    work_dir = tempfile(), timeout = 60, extra_args = c("--disable-gpu"), 
    verbose = FALSE)
}
\arguments{
\item{url}{A URL or local file path to a web page.}

\item{output}{The (PDF) output filename. For a local web page
\file{foo/bar.html}, the default PDF output is \file{foo/bar.pdf}; for a
remote URL \file{https://www.example.org/foo/bar.html}, the default output
will be \file{bar.pdf} under the current working directory.}

\item{browser}{Path to Google Chrome or Chromium. This function will try to
find it automatically via \code{\link{find_chrome}()} if the path is not
explicitly provided.}

\item{work_dir}{Name of headless Chrome working directory. If the default
temporary directory doesn't work, you may try to use a subdirectory of your
home directory.}

\item{timeout}{The number of seconds before canceling the document
generation. Use a larger value if the document takes longer to build.}

\item{extra_args}{Extra command-line arguments to be passed to Chrome.}

\item{verbose}{Whether to show verbose websocket connexion to headless
Chrome.}
}
\value{
Path of the output file (invisibly).
}
\description{
This is a wrapper function to execute the command \command{chrome --headless
--print-to-pdf url}. Google Chrome (or Chromium on Linux) must be installed
prior to using this function.
}
\references{
\url{https://developers.google.com/web/updates/2017/04/headless-chrome}
}
