% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\name{html_resume}
\alias{html_resume}
\title{Create a resume in HTML}
\usage{
html_resume(..., css = "resume", template = pkg_resource("html", "resume.html"), 
    number_sections = FALSE)
}
\arguments{
\item{..., css, template, number_sections}{See \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is based on Min-Zhong Lu's HTML/CSS in the Github repo
\url{https://github.com/mnjul/html-resume}. See
\url{https://pagedown.rbind.io/html-resume/} for an example.
}
