% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{jss_paged}
\alias{jss_paged}
\title{Create an article for the Journal of Statistical Software}
\usage{
jss_paged(
  ...,
  css = c("jss-fonts", "jss-page", "jss"),
  template = pkg_resource("html", "jss_paged.html"),
  csl = pkg_resource("csl", "journal-of-statistical-software.csl"),
  highlight = NULL,
  pandoc_args = NULL
)
}
\arguments{
\item{..., css, template, csl, highlight, pandoc_args}{Arguments passed to \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is similar to \code{\link{html_paged}}.
}
