% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paged.R
\name{thesis_paged}
\alias{thesis_paged}
\title{Create a paged HTML thesis document suitable for printing}
\usage{
thesis_paged(..., css = c("thesis"), template = pkg_resource("html", "thesis.html"))
}
\arguments{
\item{..., css, template}{Arguments passed to \code{\link{html_paged}()}.}
}
\value{
An R Markdown output format.
}
\description{
This output format is similar to \code{\link{html_paged}}. The only
difference is in the default stylesheets and Pandoc template. See
\url{https://pagedown.rbind.io/thesis-paged/} for an example.
}
